+++ exasol/toolbox/cli.py
@@ -1,6 +1,6 @@
 from argparse import ArgumentTypeError
 
-from exasol.toolbox.release import Version
+from exasol.toolbox.version import Version
 
 
 def version(arg: str) -> Version:
