+++ test/unit/cli_test.py
@@ -3,7 +3,7 @@ from argparse import ArgumentTypeError
 import pytest
 
 from exasol.toolbox.cli import version
-from exasol.toolbox.release import Version
+from exasol.toolbox.version import Version
 
 
 @pytest.mark.parametrize(
